pyproject.toml
[project]
name = "reflex-docs-mcp"
version = "0.1.0"
description = "MCP server providing structured access to Reflex documentation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.14.1",
    "aiosqlite>=0.19.0",
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "groq>=1.0.0",
    "python-dotenv>=1.2.1",
    "mcp>=1.25.0",
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
]

[project.scripts]
reflex-docs-mcp = "reflex_docs_mcp.server:main"
index-docs = "reflex_docs_mcp.indexer:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/reflex_docs_mcp"]
